requirements.txt
astroid==3.3.5
autopep8==2.3.1
blinker==1.9.0
click==8.1.7
dill==0.3.9
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
gunicorn==23.0.0
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
mccabe==0.7.0
packaging==24.2
platformdirs==4.3.6
psycopg2-binary==2.9.10
pycodestyle==2.12.1
pylint==3.3.1
python-dotenv==1.0.1
SQLAlchemy==2.0.36
tomli==2.1.0
tomlkit==0.13.2
typing_extensions==4.12.2
Werkzeug==3.1.3
